ci: refactor image builds and tags for clarity
